해시라우트 및 프로젝트 구조 변경
