+++ src/pages/LoginPage.js
@@ -1,0 +1,21 @@
+// 로그인
+const LoginPage = () => {
+  return `
+    <main class="bg-gray-100 flex items-center justify-center min-h-screen">
+        <div class="bg-white p-8 rounded-lg shadow-md w-full max-w-md">
+        <h1 class="text-2xl font-bold text-center text-blue-600 mb-8">항해플러스</h1>
+        <form id="login-form">
+            <div class="mb-4">
+            <input type="text" id="username" placeholder="사용자 이름" class="w-full p-2 border rounded" required>
+            </div>
+            <div class="mb-4">
+            <textarea id="bio" placeholder="간단한 소개" class="w-full p-2 border rounded h-24" required></textarea>
+            </div>
+            <button type="submit" class="w-full bg-blue-600 text-white p-2 rounded font-bold">로그인</button>
+        </form>
+        </div>
+    </main>
+    `;
+};
+
+export default LoginPage;
